Factor out spring simulation specific behavior and created place user tool.
