+++ src/ui/sidebar/index.html
@@ -126,13 +126,13 @@
       <img src="../trussfab-globals/assets/icons/spring.png" width="54px" alt="spring">
       <div>spring</div>
     </button>
-    <button type="button" class="tool" id="spring_animation_tool">
+    <button type="button" class="tool" id="place_user_tool">
       <img src="../trussfab-globals/assets/icons/sitting_child.png" width="54px" alt="spring">
       <div>place user</div>
     </button>
-    <button type="button" class="tool" id="spring_debug_tool">
+    <button type="button" class="tool" id="demonstrate_amplitude_tool">
       <img src="../trussfab-globals/assets/icons/spring.png" width="54px" alt="spring">
-      <div>debug</div>
+      <div>demonstrate amplitude</div>
     </button>
     <!-- <button type="button" class="tool" id="pose_check_tool">
             <img src="../trussfab-globals/assets/icons/act_dashed.png" alt="pose check">
@@ -189,10 +189,6 @@
     <button type="button" class="tool" id="add_weight_tool">
       <img src="../trussfab-globals/assets/icons/force.png" alt="add test load">
       <div>weight</div>
-    </button>
-    <button type="button" class="tool" id="place_user_tool">
-      <img src="../trussfab-globals/assets/icons/sitting_child.png" alt="add test load">
-      <div>place user</div>
     </button>
     <!-- TODO: -->
     <button type="button" class="tool" id="add_force_tool">
